Mejoras en la aplicación:
+ Se separan en casos de usos las validaciones para el stock del producto y obtener el precio total del movimiento de inventario.

+++ src/KardexMovement/Application/UseCases/DeleteKardexMovementUseCase.php
@@ -2,19 +2,29 @@
 
 namespace Src\KardexMovement\Application\UseCases;
 
+use App\Exceptions\CustomJsonException;
 use Src\KardexMovement\Domain\Repositories\KardexMovementRepositoryInterface;
 
 class DeleteKardexMovementUseCase
 {
     private $kardexMovementRepository;
+    private $getKardexMovementByIdUseCase;
 
-    public function __construct(KardexMovementRepositoryInterface $kardexMovementRepository)
+    public function __construct(
+        GetKardexMovementByIdUseCase $getKardexMovementByIdUseCase,
+        KardexMovementRepositoryInterface $kardexMovementRepository
+    )
     {
         $this->kardexMovementRepository = $kardexMovementRepository;
+        $this->getKardexMovementByIdUseCase = $getKardexMovementByIdUseCase;
     }
 
     public function execute(int $id)
     {
+        $movement = $this->getKardexMovementByIdUseCase->execute($id);
+        if (!$movement) {
+            throw new CustomJsonException(['status' => 'info', 'message' => 'Kardex movement not found!', 'code' => 404]);
+        }
         return $this->kardexMovementRepository->delete($id);
     }
 }
